bot/requirements.txt
python-telegram-bot

pandas

SQLAlchemy

psycopg2-binary

pika

openpyxl

xlrd

python-dotenv
psycopg